# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phantom Mask API",
        "description": "API for managing pharmacies, masks, and purchases",
        "contact": {
            "email": "rouyucc@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pharmacies": {
            "get": {
                "tags": [
                    "Pharmacies"
                ],
                "summary": "列出特定時間和星期幾營業的藥局",
                "description": "回傳在指定日期和時間營業的藥局列表。",
                "operationId": "getPharmacyByOpenHour",
                "parameters": [
                    {
                        "name": "day",
                        "in": "query",
                        "description": "星期幾（1-7，其中 1 代表星期一）",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "maximum": 7,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "time",
                        "in": "query",
                        "description": "24 小時制時間格式（HH:MM）",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "藥局列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "藥局 ID",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "description": "藥局名稱",
                                                "type": "string",
                                                "example": "健康藥局"
                                            },
                                            "day_of_week": {
                                                "description": "營業日",
                                                "type": "string",
                                                "example": "Monday"
                                            },
                                            "open_time": {
                                                "description": "開始營業時間",
                                                "type": "string",
                                                "example": "09:00"
                                            },
                                            "close_time": {
                                                "description": "結束營業時間",
                                                "type": "string",
                                                "example": "18:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "無效的參數請求",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "無效的參數請求"
                                        },
                                        "errors": {
                                            "properties": {
                                                "day": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "日期欄位必填"
                                                    }
                                                },
                                                "time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "時間欄位必填"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pharmacies/{id}/masks": {
            "get": {
                "tags": [
                    "Pharmacies"
                ],
                "summary": "列出特定藥局販售的所有口罩",
                "description": "回傳指定藥局及其銷售的口罩列表，口罩列表依價格由小至大排序。",
                "operationId": "getPharmacyMasks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "藥局 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "口罩列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pharmacy_id": {
                                            "description": "藥局 ID",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "pharmacy_name": {
                                            "description": "藥局名稱",
                                            "type": "string",
                                            "example": "健康藥局"
                                        },
                                        "masks": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Mask"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "無效的參數請求",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "無效的參數請求"
                                        },
                                        "errors": {
                                            "properties": {
                                                "sort": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "排序參數必須是 price 或 name"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到藥局",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "找不到指定的藥局"
                                        },
                                        "requested_id": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pharmacies/filter": {
            "get": {
                "tags": [
                    "Pharmacies"
                ],
                "summary": "列出特定價格範圍內有超過或少於 x 種口罩產品的藥局",
                "description": "根據藥局販售的口罩種類數量和價格範圍來過濾藥局，回傳符合條件的藥局及其銷售的口罩詳細資料。",
                "operationId": "getPharmaciesByMaskFilter",
                "parameters": [
                    {
                        "name": "mask_count",
                        "in": "query",
                        "description": "口罩數量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "comparison",
                        "in": "query",
                        "description": "比較方式：超過(more)或少於(less)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "more",
                                "less"
                            ]
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "最低價格",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "最高價格",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0.01
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帶有口罩詳細資料的藥局列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "藥局 ID",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "description": "藥局名稱",
                                                "type": "string",
                                                "example": "健康藥局"
                                            },
                                            "mask_count": {
                                                "description": "符合過濾條件的口罩數量",
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "masks": {
                                                "description": "符合條件的口罩列表",
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "口罩 ID",
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "brand": {
                                                            "description": "口罩品牌",
                                                            "type": "string",
                                                            "example": "3M"
                                                        },
                                                        "color": {
                                                            "description": "口罩顏色",
                                                            "type": "string",
                                                            "example": "白色"
                                                        },
                                                        "quantity_per_pack": {
                                                            "description": "每包口罩數量",
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "price": {
                                                            "description": "口罩價格",
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 25.5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "無效的參數請求",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "無效的參數請求"
                                        },
                                        "errors": {
                                            "properties": {
                                                "mask_count": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "口罩數量必須是整數"
                                                    }
                                                },
                                                "comparison": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "比較方式必須是：超過(more)或少於(less)"
                                                    }
                                                },
                                                "min_price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "最低價格必須是正數"
                                                    }
                                                },
                                                "max_price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "最高價格必須是正數"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Mask": {
                "title": "Mask",
                "description": "口罩",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand": {
                        "description": "口罩品牌",
                        "type": "string",
                        "example": "True Barrier"
                    },
                    "color": {
                        "description": "口罩顏色",
                        "type": "string",
                        "example": "blue"
                    },
                    "quantity_per_pack": {
                        "description": "每包口罩數量",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}